+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -28,10 +28,6 @@
 #import "NSData+Unzip.h"
 #import "NSMutableDictionary+Sentry.h"
 #import "Sentry/Sentry-Swift.h"
-#if !SDK_V10
-#    import "SentryANRTrackerV1.h"
-#    import "SentryANRTrackerV2.h"
-#endif
 #import "SentryAppStartMeasurement+Private.h"
 #import "SentryAppStartMeasurementProvider.h"
 #import "SentryAttachment+Private.h"